requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens==2.4.1
attrs==23.1.0
backcall==0.2.0
backports==1.0
backports.functools_lru_cache==1.6.5
bzip2==1.0.8
ca-certificates==2023.7.22
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.1
fonttools==4.43.1
fsspec==2023.10.0
gensim==4.3.2
idna==3.4
importlib-metadata==6.8.0
importlib_metadata==6.8.0
ipykernel==6.26.0
ipython==8.16.1
iso639-lang==2.1.0
jedi==0.19.1
jieba==0.42.1
jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.5.0
jupyter_core==5.5.0
kiwisolver==1.4.5
langconv==0.2.0
libffi==3.4.2
libsodium==1.0.18
libsqlite==3.43.2
libzlib==1.2.13
lightning-utilities==0.9.0
markupsafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
openssl==3.1.4
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
pillow==10.1.0
pip==23.3.1
platformdirs==3.11.0
portalocker==2.8.2
prompt-toolkit==3.0.39
prompt_toolkit==3.0.39
psutil==5.9.5
pure_eval==0.2.2
pygments==2.16.1
pyparsing==3.1.1
python==3.10.13
python-dateutil==2.8.2
python_abi==3.10
pywin32==306
pyyaml==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
scipy==1.11.3
setuptools==68.2.2
six==1.16.0
smart-open==6.4.0
stack_data==0.6.2
sympy==1.12
tk==8.6.13
torch==2.1.0
torchaudio==2.1.0
torchdata==0.7.0
torchmetrics==1.2.0
torchtext==0.16.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing-extensions==4.8.0
typing_extensions==4.8.0
tzdata==2023c
ucrt==10.0.22621.0
urllib3==2.0.7
vc==14.3
vc14_runtime==14.36.32532
vs2015_runtime==14.36.32532
wcwidth==0.2.9
wheel==0.41.3
xz==5.2.6
zeromq==4.3.5
zhconv==1.4.3
zipp==3.17.0
